requirements.txt
Django
Babel
django-phonenumber-field
phonenumberslite
djangorestframework
pip
setuptools
wheel
django-allauth
django-countries
